libvirt_requirements.txt
libvirt-python==9.1.0
